pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "threaded-earth"
version = "0.1.0"
description = "Observable local-first synthetic society simulation foundation."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "pydantic>=2.7",
  "pyyaml>=6.0",
  "sqlalchemy>=2.0",
  "typer>=0.12",
  "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]

[project.scripts]
threaded-earth = "threaded_earth.cli:app"

[tool.setuptools.packages.find]
include = ["threaded_earth*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
